refactor(cli): move command declarations to public modules
