+++ airbyte/cli/cloud/connections.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte cloud connections` CLI reference.
+"""`airbyte cloud connections` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,196 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte._util import api_util
+from airbyte.cli.cloud._api_helpers import connection_to_dict, get_auth_context, job_to_dict
+from airbyte.cli.cloud._json_helpers import (
+    JsonHelpCommand,
+    JsonHelpGroup,
+    error_json,
+    json_output,
+    parse_json_option,
+    register_schema,
+)
+
+
+@click.group(cls=JsonHelpGroup)
+@click.pass_context
+def connections(ctx: click.Context) -> None:
+    """Manage Airbyte connections."""
+    pass
+
+
+register_schema(
+    "connections_list",
+    description="List all connections in the workspace.",
+    required_params={"workspace_id": "The workspace ID."},
+)
+
+
+@connections.command("list", cls=JsonHelpCommand)
+@click.pass_context
+def connections_list(ctx: click.Context) -> None:
+    """List connections in the workspace."""
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    results = api_util.list_connections(
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output([connection_to_dict(c) for c in results])
+
+
+register_schema(
+    "connections_get",
+    description="Get details of a specific connection.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "connection_id": "The connection ID to retrieve.",
+    },
+)
+
+
+@connections.command("get", cls=JsonHelpCommand)
+@click.option("--connection-id", default=None, help="The connection ID to retrieve.")
+@click.pass_context
+def connections_get(ctx: click.Context, connection_id: str | None) -> None:
+    """Get details of a specific connection."""
+    if not connection_id:
+        error_json("--connection-id is required.", type="MissingParameter")
+    assert connection_id is not None
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    result = api_util.get_connection(
+        workspace_id=workspace_id,
+        connection_id=connection_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(connection_to_dict(result))
+
+
+register_schema(
+    "connections_create",
+    description="Create a new connection between a source and destination.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "name": "Display name for the connection.",
+        "source_id": "The source ID.",
+        "destination_id": "The destination ID.",
+    },
+    optional_params={
+        "selected_stream_names": "List of stream names to sync.",
+        "prefix": "Optional table prefix for destination.",
+    },
+)
+
+
+@connections.command("create", cls=JsonHelpCommand)
+@click.option(
+    "--json",
+    "json_str",
+    default=None,
+    help='JSON config: {"name": "...", "source_id": "...", "destination_id": "...", ...}',
+)
+@click.pass_context
+def connections_create(ctx: click.Context, json_str: str | None) -> None:
+    """Create a new connection."""
+    if not json_str:
+        error_json("--json is required.", type="MissingParameter")
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    config = parse_json_option(json_str)
+    name = config.get("name")
+    source_id = config.get("source_id")
+    destination_id = config.get("destination_id")
+    if not name or not source_id or not destination_id:
+        error_json("'name', 'source_id', and 'destination_id' are required in --json config.")
+    selected_streams: list[str] = config.get("selected_stream_names", [])
+    prefix: str = config.get("prefix", "")
+    result = api_util.create_connection(
+        name=str(name),
+        source_id=str(source_id),
+        destination_id=str(destination_id),
+        workspace_id=workspace_id,
+        prefix=prefix,
+        selected_stream_names=selected_streams,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(connection_to_dict(result))
+
+
+register_schema(
+    "connections_delete",
+    description="Delete a connection by ID.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "connection_id": "The connection ID to delete.",
+    },
+    optional_params={"force": "Skip delete safety checks (default: false)."},
+)
+
+
+@connections.command("delete", cls=JsonHelpCommand)
+@click.option("--connection-id", default=None, help="The connection ID to delete.")
+@click.option("--force", is_flag=True, default=False, help="Skip delete safety checks.")
+@click.pass_context
+def connections_delete(
+    ctx: click.Context,
+    connection_id: str | None,
+    force: bool,  # noqa: FBT001
+) -> None:
+    """Delete a connection."""
+    if not connection_id:
+        error_json("--connection-id is required.", type="MissingParameter")
+    assert connection_id is not None
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    api_util.delete_connection(
+        connection_id,
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+        safe_mode=not force,
+    )
+    json_output({"deleted": True, "connection_id": connection_id})
+
+
+register_schema(
+    "connections_sync",
+    description="Trigger a sync job for a connection.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "connection_id": "The connection ID to sync.",
+    },
+)
+
+
+@connections.command("sync", cls=JsonHelpCommand)
+@click.option("--connection-id", default=None, help="The connection ID to sync.")
+@click.pass_context
+def connections_sync(ctx: click.Context, connection_id: str | None) -> None:
+    """Trigger a sync for a connection."""
+    if not connection_id:
+        error_json("--connection-id is required.", type="MissingParameter")
+    assert connection_id is not None
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    result = api_util.run_connection(
+        workspace_id,
+        connection_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(job_to_dict(result))
+
+
+__all__ = ["connections"]
