+++ airbyte/cli/cloud/destinations.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte cloud destinations` CLI reference.
+"""`airbyte cloud destinations` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,155 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte._util import api_util
+from airbyte.cli.cloud._api_helpers import (
+    destination_to_dict,
+    get_auth_context,
+    get_auth_no_workspace,
+)
+from airbyte.cli.cloud._json_helpers import (
+    JsonHelpCommand,
+    JsonHelpGroup,
+    error_json,
+    json_output,
+    parse_json_option,
+    register_schema,
+)
+
+
+@click.group(cls=JsonHelpGroup)
+@click.pass_context
+def destinations(ctx: click.Context) -> None:
+    """Manage Airbyte destinations."""
+    pass
+
+
+register_schema(
+    "destinations_list",
+    description="List all destinations in the workspace.",
+    required_params={"workspace_id": "The workspace ID."},
+)
+
+
+@destinations.command("list", cls=JsonHelpCommand)
+@click.pass_context
+def destinations_list(ctx: click.Context) -> None:
+    """List destinations in the workspace."""
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    results = api_util.list_destinations(
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output([destination_to_dict(d) for d in results])
+
+
+register_schema(
+    "destinations_get",
+    description="Get details of a specific destination.",
+    required_params={"destination_id": "The destination ID to retrieve."},
+)
+
+
+@destinations.command("get", cls=JsonHelpCommand)
+@click.option("--destination-id", default=None, help="The destination ID to retrieve.")
+@click.pass_context
+def destinations_get(ctx: click.Context, destination_id: str | None) -> None:
+    """Get details of a specific destination."""
+    if not destination_id:
+        error_json("--destination-id is required.", type="MissingParameter")
+    assert destination_id is not None
+    api_url, client_id, client_secret = get_auth_no_workspace(ctx)
+    result = api_util.get_destination(
+        destination_id=destination_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(destination_to_dict(result))
+
+
+register_schema(
+    "destinations_create",
+    description="Create a new destination in the workspace.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "name": "Display name for the destination.",
+        "destinationType": "The destination connector type (e.g. 'bigquery').",
+    },
+    optional_params={"...": "Additional connector-specific configuration fields."},
+)
+
+
+@destinations.command("create", cls=JsonHelpCommand)
+@click.option(
+    "--json",
+    "json_str",
+    default=None,
+    help='JSON config: {"name": "...", "destinationType": "...", ...}',
+)
+@click.pass_context
+def destinations_create(ctx: click.Context, json_str: str | None) -> None:
+    """Create a new destination in the workspace."""
+    if not json_str:
+        error_json("--json is required.", type="MissingParameter")
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    config = parse_json_option(json_str)
+    name = config.pop("name", None)
+    if not name:
+        error_json("'name' is required in --json config.")
+    result = api_util.create_destination(
+        name=name,
+        workspace_id=workspace_id,
+        config=config,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(destination_to_dict(result))
+
+
+register_schema(
+    "destinations_delete",
+    description="Delete a destination by ID.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "destination_id": "The destination ID to delete.",
+    },
+    optional_params={"force": "Skip delete safety checks (default: false)."},
+)
+
+
+@destinations.command("delete", cls=JsonHelpCommand)
+@click.option("--destination-id", default=None, help="The destination ID to delete.")
+@click.option("--force", is_flag=True, default=False, help="Skip delete safety checks.")
+@click.pass_context
+def destinations_delete(
+    ctx: click.Context,
+    destination_id: str | None,
+    force: bool,  # noqa: FBT001
+) -> None:
+    """Delete a destination."""
+    if not destination_id:
+        error_json("--destination-id is required.", type="MissingParameter")
+    assert destination_id is not None
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    api_util.delete_destination(
+        destination_id=destination_id,
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+        safe_mode=not force,
+    )
+    json_output({"deleted": True, "destination_id": destination_id})
+
+
+__all__ = ["destinations"]
