+++ airbyte/cli/cloud/jobs.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte cloud jobs` CLI reference.
+"""`airbyte cloud jobs` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,83 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte._util import api_util
+from airbyte.cli.cloud._api_helpers import get_auth_context, get_auth_no_workspace, job_to_dict
+from airbyte.cli.cloud._json_helpers import (
+    JsonHelpCommand,
+    JsonHelpGroup,
+    error_json,
+    json_output,
+    register_schema,
+)
+
+
+@click.group(cls=JsonHelpGroup)
+@click.pass_context
+def jobs(ctx: click.Context) -> None:
+    """View Airbyte sync jobs."""
+    pass
+
+
+register_schema(
+    "jobs_list",
+    description="List recent sync jobs for a connection.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "connection_id": "The connection ID.",
+    },
+    optional_params={"limit": "Maximum number of jobs to return (default: 20)."},
+)
+
+
+@jobs.command("list", cls=JsonHelpCommand)
+@click.option("--connection-id", default=None, help="The connection ID to list jobs for.")
+@click.option("--limit", default=20, help="Maximum number of jobs to return.")
+@click.pass_context
+def jobs_list(ctx: click.Context, connection_id: str | None, limit: int) -> None:
+    """List recent jobs for a connection."""
+    if not connection_id:
+        error_json("--connection-id is required.", type="MissingParameter")
+    assert connection_id is not None
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    results = api_util.get_job_logs(
+        workspace_id,
+        connection_id,
+        limit,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output([job_to_dict(j) for j in results])
+
+
+register_schema(
+    "jobs_get",
+    description="Get details of a specific job by ID.",
+    required_params={"job_id": "The job ID to retrieve."},
+)
+
+
+@jobs.command("get", cls=JsonHelpCommand)
+@click.option("--job-id", default=None, type=int, help="The job ID to retrieve.")
+@click.pass_context
+def jobs_get(ctx: click.Context, job_id: int | None) -> None:
+    """Get details of a specific job."""
+    if job_id is None:
+        error_json("--job-id is required.", type="MissingParameter")
+    assert job_id is not None
+    api_url, client_id, client_secret = get_auth_no_workspace(ctx)
+    result = api_util.get_job_info(
+        job_id=job_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(job_to_dict(result))
+
+
+__all__ = ["jobs"]
