+++ airbyte/cli/cloud/sources.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte cloud sources` CLI reference.
+"""`airbyte cloud sources` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,147 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte._util import api_util
+from airbyte.cli.cloud._api_helpers import get_auth_context, get_auth_no_workspace, source_to_dict
+from airbyte.cli.cloud._json_helpers import (
+    JsonHelpCommand,
+    JsonHelpGroup,
+    error_json,
+    json_output,
+    parse_json_option,
+    register_schema,
+)
+
+
+@click.group(cls=JsonHelpGroup)
+@click.pass_context
+def sources(ctx: click.Context) -> None:
+    """Manage Airbyte sources."""
+    pass
+
+
+register_schema(
+    "sources_list",
+    description="List all sources in the workspace.",
+    required_params={"workspace_id": "The workspace ID."},
+)
+
+
+@sources.command("list", cls=JsonHelpCommand)
+@click.pass_context
+def sources_list(ctx: click.Context) -> None:
+    """List sources in the workspace."""
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    results = api_util.list_sources(
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output([source_to_dict(s) for s in results])
+
+
+register_schema(
+    "sources_get",
+    description="Get details of a specific source.",
+    required_params={"source_id": "The source ID to retrieve."},
+)
+
+
+@sources.command("get", cls=JsonHelpCommand)
+@click.option("--source-id", default=None, help="The source ID to retrieve.")
+@click.pass_context
+def sources_get(ctx: click.Context, source_id: str | None) -> None:
+    """Get details of a specific source."""
+    if not source_id:
+        error_json("--source-id is required.", type="MissingParameter")
+    assert source_id is not None
+    api_url, client_id, client_secret = get_auth_no_workspace(ctx)
+    result = api_util.get_source(
+        source_id=source_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(source_to_dict(result))
+
+
+register_schema(
+    "sources_create",
+    description="Create a new source in the workspace.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "name": "Display name for the source.",
+        "sourceType": "The source connector type (e.g. 'postgres').",
+    },
+    optional_params={"...": "Additional connector-specific configuration fields."},
+)
+
+
+@sources.command("create", cls=JsonHelpCommand)
+@click.option(
+    "--json",
+    "json_str",
+    default=None,
+    help='JSON config: {"name": "...", "sourceType": "...", ...}',
+)
+@click.pass_context
+def sources_create(ctx: click.Context, json_str: str | None) -> None:
+    """Create a new source in the workspace."""
+    if not json_str:
+        error_json("--json is required.", type="MissingParameter")
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    config = parse_json_option(json_str)
+    name = config.pop("name", None)
+    if not name:
+        error_json("'name' is required in --json config.")
+    result = api_util.create_source(
+        name=name,
+        workspace_id=workspace_id,
+        config=config,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(source_to_dict(result))
+
+
+register_schema(
+    "sources_delete",
+    description="Delete a source by ID.",
+    required_params={
+        "workspace_id": "The workspace ID.",
+        "source_id": "The source ID to delete.",
+    },
+    optional_params={"force": "Skip delete safety checks (default: false)."},
+)
+
+
+@sources.command("delete", cls=JsonHelpCommand)
+@click.option("--source-id", default=None, help="The source ID to delete.")
+@click.option("--force", is_flag=True, default=False, help="Skip delete safety checks.")
+@click.pass_context
+def sources_delete(ctx: click.Context, source_id: str | None, force: bool) -> None:  # noqa: FBT001
+    """Delete a source."""
+    if not source_id:
+        error_json("--source-id is required.", type="MissingParameter")
+    assert source_id is not None
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    api_util.delete_source(
+        source_id=source_id,
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+        safe_mode=not force,
+    )
+    json_output({"deleted": True, "source_id": source_id})
+
+
+__all__ = ["sources"]
