+++ airbyte/cli/cloud/workspaces.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte cloud workspaces` CLI reference.
+"""`airbyte cloud workspaces` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,74 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte._util import api_util
+from airbyte.cli._cli_auth import resolve_workspace_id
+from airbyte.cli.cloud._api_helpers import (
+    get_auth_context,
+    get_auth_no_workspace,
+    workspace_to_dict,
+)
+from airbyte.cli.cloud._json_helpers import (
+    JsonHelpCommand,
+    JsonHelpGroup,
+    json_output,
+    register_schema,
+)
+
+
+@click.group(cls=JsonHelpGroup)
+@click.pass_context
+def workspaces(ctx: click.Context) -> None:
+    """Manage Airbyte workspaces."""
+    pass
+
+
+register_schema(
+    "workspaces_list",
+    description="List all workspaces accessible with the current credentials.",
+    optional_params={"workspace_id": "Filter to a specific workspace ID."},
+)
+
+
+@workspaces.command("list", cls=JsonHelpCommand)
+@click.pass_context
+def workspaces_list(ctx: click.Context) -> None:
+    """List workspaces accessible with the current credentials."""
+    api_url, client_id, client_secret, workspace_id = get_auth_context(ctx)
+    results = api_util.list_workspaces(
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output([workspace_to_dict(w) for w in results])
+
+
+register_schema(
+    "workspaces_get",
+    description="Get details of a specific workspace.",
+    required_params={"workspace_id": "The workspace ID to retrieve."},
+)
+
+
+@workspaces.command("get", cls=JsonHelpCommand)
+@click.option("--workspace-id", "cmd_workspace_id", default=None, help="Workspace ID to retrieve.")
+@click.pass_context
+def workspaces_get(ctx: click.Context, cmd_workspace_id: str | None) -> None:
+    """Get details of a specific workspace."""
+    api_url, client_id, client_secret = get_auth_no_workspace(ctx)
+    workspace_id = resolve_workspace_id(cmd_workspace_id or ctx.obj["_raw_workspace_id"])
+    result = api_util.get_workspace(
+        workspace_id=workspace_id,
+        api_root=api_url,
+        client_id=client_id,
+        client_secret=client_secret,
+        bearer_token=None,
+    )
+    json_output(workspace_to_dict(result))
+
+
+__all__ = ["workspaces"]
