+++ airbyte/cli/local/connectors.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte local connectors` CLI reference.
+"""`airbyte local connectors` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,19 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte.cli.pyab import benchmark, validate
+
+
+@click.group()
+def connectors() -> None:
+    """Develop and inspect local connector behavior."""
+    pass
+
+
+connectors.add_command(validate)
+connectors.add_command(benchmark)
+
+
+__all__ = ["connectors"]
