+++ airbyte/cli/local/debug.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte local debug` CLI reference.
+"""`airbyte local debug` commands.
 
 The commands below are regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,24 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+import click
 
-__all__: list[str] = []
+from airbyte.cli.pyab import destination_smoke_test
+
+
+@click.group()
+def debug() -> None:
+    """Debug local connector workflows."""
+    pass
+
+
+@debug.group()
+def destinations() -> None:
+    """Debug destination behavior with synthetic records."""
+    pass
+
+
+destinations.add_command(destination_smoke_test, name="smoke-test")
+
+
+__all__ = ["debug"]
