+++ airbyte/cli/local/sync.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-"""`airbyte local sync` CLI reference.
+"""`airbyte local sync` command.
 
 The command below is regenerated by `poe docs-generate` from the Click command
 tree; see `docs/generate_cli.py`.
@@ -10,5 +10,7 @@ tree; see `docs/generate_cli.py`.
 
 from __future__ import annotations
 
+from airbyte.cli.pyab import sync
 
-__all__: list[str] = []
+
+__all__ = ["sync"]
